move a ton of stuff to shared.
