+++ Content.Server/Medical/CryoPodSystem.cs
@@ -7,19 +7,15 @@ using Content.Server.NodeContainer.NodeGroups;
 using Content.Server.NodeContainer.Nodes;
 using Content.Shared.Atmos;
 using Content.Shared.Medical.Cryogenics;
-using Robust.Shared.Containers;
-using Robust.Shared.Timing;
 namespace Content.Server.Medical;
 
 public sealed partial class CryoPodSystem : SharedCryoPodSystem
 {
-    [Dependency] private readonly IGameTiming _timing = default!;
     [Dependency] private readonly AtmosphereSystem _atmosphereSystem = default!;
     [Dependency] private readonly GasCanisterSystem _gasCanisterSystem = default!;
     [Dependency] private readonly GasAnalyzerSystem _gasAnalyzerSystem = default!;
     [Dependency] private readonly HealthAnalyzerSystem _healthAnalyzerSystem = default!;
     [Dependency] private readonly NodeContainerSystem _nodeContainer = default!;
-    [Dependency] private readonly SharedUserInterfaceSystem _uiSystem = default!;
 
 
     public override void Initialize()
@@ -28,26 +24,17 @@ public sealed partial class CryoPodSystem : SharedCryoPodSystem
 
         SubscribeLocalEvent<CryoPodComponent, AtmosDeviceUpdateEvent>(OnCryoPodUpdateAtmosphere);
         SubscribeLocalEvent<CryoPodComponent, GasAnalyzerScanEvent>(OnGasAnalyzed);
-        SubscribeLocalEvent<CryoPodComponent, EntRemovedFromContainerMessage>(OnEjected);
-        SubscribeLocalEvent<CryoPodComponent, EntInsertedIntoContainerMessage>(OnBodyInserted);
-        SubscribeLocalEvent<CryoPodComponent, CryoPodUiMessage>(OnUiMessage);
-
-        Subs.BuiEvents<CryoPodComponent>(CryoPodUiKey.Key, subs =>
-        {
-            subs.Event<BoundUIOpenedEvent>(OnBoundUiOpened);
-        });
     }
 
     public override void Update(float frameTime)
     {
         base.Update(frameTime);
 
-        var curTime = _timing.CurTime;
         var query = EntityQueryEnumerator<ActiveCryoPodComponent, CryoPodComponent>();
 
         while (query.MoveNext(out var uid, out _, out var cryoPod))
         {
-            if (curTime < cryoPod.NextUiUpdateTime)
+            if (Timing.CurTime < cryoPod.NextUiUpdateTime)
                 continue;
 
             cryoPod.NextUiUpdateTime += cryoPod.UiUpdateInterval;
@@ -56,9 +43,9 @@ public sealed partial class CryoPodSystem : SharedCryoPodSystem
         }
     }
 
-    private void UpdateUi(Entity<CryoPodComponent> entity)
+    protected override void UpdateUi(Entity<CryoPodComponent> entity)
     {
-        if (!_uiSystem.IsUiOpen(entity.Owner, CryoPodUiKey.Key)
+        if (!UI.IsUiOpen(entity.Owner, CryoPodUiKey.Key)
             || !TryComp(entity, out CryoPodAirComponent? air))
             return;
 
@@ -69,7 +56,7 @@ public sealed partial class CryoPodSystem : SharedCryoPodSystem
         var health = _healthAnalyzerSystem.GetHealthAnalyzerUiState(patient);
         health.ScanMode = true;
 
-        _uiSystem.ServerSendUiMessage(
+        UI.ServerSendUiMessage(
             entity.Owner,
             CryoPodUiKey.Key,
             new CryoPodUserMessage(gasMix, health, beakerCapacity, beaker, injecting)
@@ -108,49 +95,5 @@ public sealed partial class CryoPodSystem : SharedCryoPodSystem
             portAirLocal.Volume = port.Volume;
             args.GasMixtures.Add((entity.Comp.PortName, portAirLocal));
         }
-    }
-
-    private void OnUiMessage(Entity<CryoPodComponent> cryoPod, ref CryoPodUiMessage msg)
-    {
-        switch (msg.Type)
-        {
-            case CryoPodUiMessage.MessageType.EjectPatient:
-                TryEjectBody(cryoPod.Owner, msg.Actor, cryoPod.Comp);
-                break;
-            case CryoPodUiMessage.MessageType.EjectBeaker:
-                TryEjectBeaker(cryoPod, msg.Actor);
-                break;
-            case CryoPodUiMessage.MessageType.Inject:
-                TryInject(cryoPod, msg.Quantity.GetValueOrDefault());
-                break;
-        }
-
-        UpdateUi(cryoPod);
-    }
-
-    private void OnBoundUiOpened(Entity<CryoPodComponent> cryoPod, ref BoundUIOpenedEvent args)
-    {
-        UpdateUi(cryoPod);
-    }
-
-    private void OnEjected(Entity<CryoPodComponent> cryoPod, ref EntRemovedFromContainerMessage args)
-    {
-        if (args.Container.ID == CryoPodComponent.BodyContainerName)
-        {
-            ClearInjectionBuffer(cryoPod);
-        }
-
-        UpdateUi(cryoPod);
-    }
-
-    private void OnBodyInserted(Entity<CryoPodComponent> cryoPod, ref EntInsertedIntoContainerMessage args)
-    {
-        if (args.Container.ID == CryoPodComponent.BodyContainerName)
-        {
-            _uiSystem.CloseUi(cryoPod.Owner, CryoPodUiKey.Key, args.Entity);
-            ClearInjectionBuffer(cryoPod);
-        }
-
-        UpdateUi(cryoPod);
     }
 }
